Split options out to separate module, added more documentation
