+++ src/eant2.rs
@@ -1,88 +1,8 @@
 use std::sync::Arc;
 use cge::{Network, Activation};
-use cmaes::{restart::{RestartStrategy}};
 use typed_builder::TypedBuilder;
-use crate::{FitnessFunction, generation::Generation, select, mutation::mutate, mutation_probabilities::MutationSampler};
-
-const DEFAULT_ACTIVATION:          Activation       = Activation::Sigmoid;
-const DEFAULT_POPULATION_SIZE:     usize            = 10;
-const DEFAULT_OFFSPRING_COUNT:     usize            = 4;
-const DEFAULT_SIMILAR_FITNESS:     f64              = 0.15;
-const DEFAULT_MAX_GENERATIONS:     usize            = 30;
-const DEFAULT_TERMINATING_FITNESS: f64              = 0.0;
-const DEFAULT_EANT2_TERMINATION:   EANT2Termination = EANT2Termination { fitness: 0.0, generations: 30 };
-
-/// - When should the EANT2 algorithm terminate?
-/// - - For each EANT2 iteration, when should the CMA-ES algorithm terminate?
-#[derive(TypedBuilder)]
-pub struct EANT2Termination {
-  #[builder(default = DEFAULT_TERMINATING_FITNESS)]
-  /// Fitness crossing this threshold will cause the algorithm to terminate (think: goal reached).
-  /// - Defaults to `0.0`.
-  pub fitness: f64,
-  
-  #[builder(default = DEFAULT_MAX_GENERATIONS)]
-  /// Limit on the number of generations (iterations of the EANT2 algorithm). 
-  /// - Defaults to `30`.
-  pub generations: usize
-}
-
-#[derive(TypedBuilder)]
-pub struct Exploration {
-  #[builder(
-    default = DEFAULT_POPULATION_SIZE, 
-    setter(doc = "Sets the population size. Increasing this option may produce higher quality neural
-                  networks, but will increase the time needed to find a solution."))]
-  pub population: usize,
-
-  #[builder(
-    default = DEFAULT_OFFSPRING_COUNT, 
-    setter(doc = "Sets the offspring count (how many offspring each individual spawns). Increasing this
-                  option may produce higher quality neural networks, but will increase the time needed to
-                  find a solution."))]
-  pub offspring: usize,
-
-  #[builder(
-    default = DEFAULT_SIMILAR_FITNESS, 
-    setter(doc= "Sets the threshold for deciding whether two neural networks have a similar fitness.
-                 Increasing this option will make the algorithm more aggresively prefer smaller
-                 neural networks. Decreasing it will do the opposite, allowing larger individuals to stay in
-                 the population. It is recommended to set this option higher if a small neural network is
-                 preferred. The downside is it will take slightly longer to find a solution, due to more
-                 higher fitness neural networks being discarded."))]
-  pub similar_fitness: f64,
-
-  #[builder(
-    default_code = "MutationSampler::default()", 
-    setter(doc = "Sets the sampler which chooses each mutation.  Build a `MutationSampler` with `MutationProbabilities`."))]
-  pub mutation_probabilities: MutationSampler,
-
-  #[builder(default = DEFAULT_EANT2_TERMINATION, setter(doc = "Termination conditions (target fitness, max generations)"))]
-  pub terminate: EANT2Termination,
-}
-
-#[derive(TypedBuilder)]
-pub struct CMAESTermination {
-  #[builder(default = None, setter(strip_option, doc = "Default is to use the fitness from `EANT2Termination` options, which defaults to `0.0`."))]
-  pub fitness:     Option<f64>,
-
-  #[builder(default = None, setter(strip_option, doc = "Default is no CMA-ES generation limit."))]
-  pub generations: Option<usize>
-}
-
-#[derive(TypedBuilder)]
-pub struct Exploitation {
-  /// CMA-ES parameter optimization restart strategy. Defaults to `RestartStrategy::BIPOP`.
-  #[builder(
-    default_code = "RestartStrategy::BIPOP(Default::default())",
-    setter(doc = "CMA-ES parameter optimization (exploitation) restart strategy. Defaults to `RestartStrategy::BIPOP`."))]
-  pub restart: RestartStrategy,
-
-  #[builder(
-    default_code = "CMAESTermination::builder().build()",
-    setter(doc = "CMA-ES parameter optimization (exploitation) termination conditions (target fitness, max generations)."))]
-  pub terminate: CMAESTermination
-}
+use crate::{FitnessFunction, generation::Generation, select, mutation::mutate};
+use crate::options::*;
 
 /// The EANT2 algorithm.
 /// 
@@ -104,8 +24,6 @@ pub struct Exploitation {
 /// ```
 /// 
 /// # Minimal Example
-/// - Most options have good default values, and exist for flexibility.
-/// - See the `Advanced Example` to see all the options that are normally handled for you.
 /// 
 /// ```rust
 /// use eant2::*;
@@ -114,10 +32,12 @@ pub struct Exploitation {
 ///   .outputs(3)
 ///   .build();
 /// 
-/// let (network, fitness) = train.run();
+/// let (network, fitness) = train.run(MyFitnessFunction);
 /// ```
 /// 
 /// # Advanced Example
+/// 
+/// - Most options have good default values, and exist only for flexibility.
 /// 
 /// ```rust
 /// use eant2::*;
@@ -137,7 +57,7 @@ pub struct Exploitation {
 ///            .fitness(0.15)       // either terminate EANT2 when fitness hits 0.15,
 ///            .generations(12)     // or terminate after 12 generations
 ///       )
-///       .mutation_probabilities(
+///       .mutation_probabilities(  // describe relative mutation probabilities
 ///         MutationProbabilities::zeros()
 ///           .add_connection(2.)    
 ///           .remove_connection(2.)
@@ -159,7 +79,6 @@ pub struct Exploitation {
 ///    )
 ///   .build();
 /// 
-/// // ...
 /// ```
 #[derive(TypedBuilder)]
 #[builder(doc)]
