+++ src/lib.rs
@@ -49,8 +49,9 @@ mod mutation;
 mod generation;
 mod cge_utils;
 mod select;
-pub mod fitness;
+pub mod options;
 pub mod eant2;
+pub mod fitness;
 pub mod mutation_probabilities;
 pub use cge::{Network, Activation};
 pub use fitness::FitnessFunction;
